+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Mon, 15 May 2023 14:59:04 GMT and should not be manually modified.
+This log was last generated on Wed, 17 May 2023 21:33:40 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 11.8.30
+
+Wed, 17 May 2023 21:33:40 GMT
+
+### Patches
+
+- Add support for the column-configuration-change event to the table ([ni/nimble@71351b1](https://github.com/ni/nimble/commit/71351b15e45f68adb8b08db2168b5faf499a49f9))
 
 ## 11.8.29